requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.3
fastapi==0.101.0
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mediapipe==0.10.3
mtcnn==0.1.1
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
pydentic==0.0.1.dev3
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
python-stdnum==1.19
pytz==2023.3.post1
PyWavelets==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.8.30
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
Werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
